CEDS-1249 Move mocks to spearate package and files
